Create Random Color
React page that has default background color. Background color changes to random color with click of button. 
Text and button will change color to match that of the background as well. 
isLight() will detect whether the background color is dark or light and change font accordingly. This is also used to set the className of the button to either "light-button" or "dark-button" which will alter the background color of the button itself so it can be seen on the new background